requirements.txt

# Airflow (if running locally)
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0

# Code formatting
black==23.7.0

# dbt
dbt-core==1.6.0
dbt-postgres==1.6.0
flake8==6.0.0
matplotlib==3.7.2
numpy==1.24.3
# Core data processing
pandas==2.0.3

# Data visualization (optional)
plotly==5.15.0

# Code formatting
pre-commit==3.5.0

# Database connectivity
psycopg2-binary==2.9.7

# Development tools
pytest==7.4.0
python-dotenv==1.0.0

# Utility libraries
requests==2.31.0
sqlalchemy==2.0.20

dotenv==0.9.9